refactor: delete callback
refactor: move DayOfWeek to the separate component
