+++ src/components/NewEvent/EventRepeat/EventRepeatDays/DayOfWeek/DayOfWeek.module.css
@@ -1,0 +1,21 @@
+
+.repeat_days {
+    margin-top: 15px;
+    display: flex;
+    gap: 10px;
+    font-size: 10px;
+    margin-bottom: 30px;
+}
+
+.repeat_day {
+    background-color: rgba(100,100,100,0.1);
+    color: grey;
+    padding: 5px;
+    border-radius: 100%;
+    cursor: pointer;
+}
+
+.active {
+    background-color: #4B84F4;
+    color: white;
+}
